added test schemes and modified travis.yml
