+++ NetworkKit.xcodeproj/xcuserdata/alextelek.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -19,6 +19,11 @@
 			<key>orderHint</key>
 			<integer>3</integer>
 		</dict>
+		<key>NetworkKitMacTests.xcscheme</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>6</integer>
+		</dict>
 		<key>NetworkKitTests.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
@@ -28,6 +33,11 @@
 		<dict>
 			<key>orderHint</key>
 			<integer>5</integer>
+		</dict>
+		<key>NetworkKitTvOSTests.xcscheme</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>7</integer>
 		</dict>
 		<key>NetworkKitWatchOS.xcscheme_^#shared#^_</key>
 		<dict>
@@ -48,6 +58,16 @@
 			<true/>
 		</dict>
 		<key>2BB9D8591C9AF95000A9F47B</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>2BB9D8851C9F1A2800A9F47B</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>2BB9D8951C9F1A4900A9F47B</key>
 		<dict>
 			<key>primary</key>
 			<true/>
