+++ NetworkKit/Info.plist
@@ -19,12 +19,7 @@
 	<key>CFBundleSignature</key>
 	<string>????</string>
 	<key>CFBundleVersion</key>
-	<string>166</string>
-	<key>NSAppTransportSecurity</key>
-	<dict>
-		<key>NSAllowsArbitraryLoads</key>
-		<true/>
-	</dict>
+	<string>169</string>
 	<key>NSPrincipalClass</key>
 	<string></string>
 </dict>
