+++ NetworkKitMac/Info.plist
@@ -20,19 +20,6 @@
 	<string>????</string>
 	<key>CFBundleVersion</key>
 	<string>16</string>
-	<key>NSAppTransportSecurity</key>
-	<dict>
-		<key>NSAllowsArbitraryLoads</key>
-		<true/>
-		<key>NSExceptionDomains</key>
-		<dict>
-			<key>https://hacker-news.firebaseio.com</key>
-			<dict>
-				<key>NSIncludesSubdomains</key>
-				<true/>
-			</dict>
-		</dict>
-	</dict>
 	<key>NSHumanReadableCopyright</key>
 	<string>Copyright © 2016 alextelek. All rights reserved.</string>
 	<key>NSPrincipalClass</key>
