+++ NetworkKitMacTests/Info.plist
